src/main/java/com/sapling/common/tools/common/ChineseCharUtil.java
package com.sapling.common.tools.common;

/**
 * @author weizhou
 * @version v1.0
 * @date 2018/7/2
 * @since v1.0
 */
public class ChineseCharUtil {

    public static boolean isChineseChar(char c){
        Character.UnicodeBlock ub = Character.UnicodeBlock.of(c);
        if (ub == Character.UnicodeBlock.CJK_UNIFIED_IDEOGRAPHS
                || ub == Character.UnicodeBlock.CJK_COMPATIBILITY_IDEOGRAPHS
                || ub == Character.UnicodeBlock.CJK_UNIFIED_IDEOGRAPHS_EXTENSION_A
                || ub == Character.UnicodeBlock.GENERAL_PUNCTUATION
                || ub == Character.UnicodeBlock.CJK_SYMBOLS_AND_PUNCTUATION
                || ub == Character.UnicodeBlock.HALFWIDTH_AND_FULLWIDTH_FORMS) {
            return true;
        }
        return false;
    }

    public static boolean includeChinese(String value){
        if (StringUtil.isEmpty(value)){
            return false;
        }
        char[] chars = value.toCharArray();
        for (char c : chars){
            if (isChineseChar(c)){
                return true;
            }
        }
        return false;
    }

}
